requirements.txt
diff-match-patch==20121119
Django==1.7.8
django-ckeditor==4.5.1
django-flat-theme==0.9.3
django-import-export==0.2.8
django-multiselectfield==0.1.3
django-smart-selects==1.0.9.3
Pillow==3.0.0
psycopg2==2.6.1
sorl-thumbnail==12.3
tablib==0.10.0
